requirements.txt
ultralytics==8.0.227
opencv-python-headless==4.9.0.80
numpy==1.26.2
Pillow==10.1.0
streamlit==1.29.0
streamlit-webrtc==0.47.1